backend/requirements.txt
asgiref==3.8.1
boto3==1.26.27
botocore==1.29.165
Django==3.2.25
django-cors-headers==3.7.0
djangorestframework==3.12.4
jmespath==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.6.2
six==1.16.0
sqlparse==0.5.0
urllib3==1.26.18
